+++ ui/src/layout/PlaylistSubmenu.js
@@ -1,22 +1,19 @@
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
-import { useRefresh, MenuItemLink, useGetList, useTranslate } from 'react-admin'
+import { useRefresh, MenuItemLink, useGetList } from 'react-admin'
 import Playlist from '../icons/Playlist'
 import SubMenu from './SubMenu'
-import { translatedResourceName } from '../utils'
 
 const PlaylistSubmenu = ({
   isSidebarOpen,
   isToggled,
-  resources,
+  name,
   dense,
   handleToggle,
   onMenuClick,
 }) => {
   const refresh = useRefresh()
-  const translate = useTranslate()
   const [playLists, setPlaylists] = useState([])
-  const [menuName, setMenuName] = useState('')
   const { data } = useGetList('playlist', { page: 1, perPage: -1 }, {}, {})
 
   useEffect(() => {
@@ -29,16 +26,6 @@ const PlaylistSubmenu = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [data])
 
-  useEffect(() => {
-    if (resources?.length) {
-      const playListResource = resources.find(
-        (resource) => resource.name === 'playlist'
-      )
-      setMenuName(translatedResourceName(playListResource, translate))
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [resources])
-
   return (
     <div
       style={{
@@ -50,7 +37,7 @@ const PlaylistSubmenu = ({
         handleToggle={() => handleToggle('menuPlaylists')}
         isOpen={isToggled}
         sidebarIsOpen={isSidebarOpen}
-        name={menuName}
+        name={name}
         icon={<Playlist />}
         dense={dense}
         secondaryAction={onMenuClick}
@@ -77,7 +64,7 @@ export default PlaylistSubmenu
 PlaylistSubmenu.propTypes = {
   isSidebarOpen: PropTypes.bool.isRequired,
   isToggled: PropTypes.bool.isRequired,
-  resources: PropTypes.array.isRequired,
+  name: PropTypes.string.isRequired,
   handleToggle: PropTypes.func.isRequired,
   onMenuClick: PropTypes.func.isRequired,
   dense: PropTypes.bool,
