+++ ui/src/utils/index.js
@@ -2,4 +2,3 @@ export * from './baseUrl'
 export * from './docsUrl'
 export * from './formatters'
 export * from './notifications'
-export * from './translateResource'
